+++ domain/usecase/src/main/java/co/com/bancolombia/usecase/exceptions/MontoFueraDeRangoException.java
@@ -1,12 +1,13 @@
 package co.com.bancolombia.usecase.exceptions;
 
+import java.math.BigDecimal;
 import java.text.DecimalFormat;
 
 public class MontoFueraDeRangoException extends RuntimeException {
 
     private static final DecimalFormat formatter = new DecimalFormat("#,###.##");
 
-    public MontoFueraDeRangoException(Double monto, Double minimo, Double maximo) {
+    public MontoFueraDeRangoException(BigDecimal monto, BigDecimal minimo, BigDecimal maximo) {
         super("Monto " + formatter.format(monto)
                 + " fuera del rango permitido [" + formatter.format(minimo)
                 + " - " + formatter.format(maximo) + "]");
